api/requirements.txt
appnope==0.1.0
backcall==0.1.0
boto3==1.9.33
botocore==1.12.33
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-rest-swagger==2.1.2
itypes==1.1.0
openapi-codec==1.3.2
simplejson==3.16.0
uritemplate==3.0.0
decorator==4.2.1
disposable-email-domains==0.0.46
django==2.0.3
django-cors-headers==2.2.0
django-heroku==0.3.1
django-extensions==2.1.2
django-filter==1.1.0
django-model-utils==3.1.2
django-storages==1.7.1
django-redis-cache==1.8.1
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
dj-database-url==0.5.0
Faker==0.9.1
factory-boy==2.11.1
flake8==3.6.0
honcho==1.0.1
httpie==0.9.9
idna==2.6
ipython==6.3.0
ipython-genutils==0.2.0
jedi==0.11.1
mailjet-rest==1.3.0
Markdown==2.6.11
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.3
pylint-django==2.0.5
pylint-plugin-utils==0.4
redis==2.10.6
requests==2.18.4
sentry-sdk==0.4.1
simplegeneric==0.8.1
six==1.11.0
tblib==1.3.2
traitlets==4.3.2
text-unidecode==1.2
urllib3==1.22
wcwidth==0.1.7
whitenoise==3.3.1
